Fix schedules not switching for output
The issue: although you're making a copy of Schedule's timetable,
you're not copying its constraints. This messes up the SoftCount,
giving the illusion of the same schedule. In fact, the schedules
are indeed different. So here, instead of using the SoftCount
of the schedule, we simply use the SoftCount stored in Solver's
map, which is the same.